requirements.txt
flask
gunicorn
opencv-python-headless
numpy
tensorflow-cpu
keras
pillow
